Move model files to models folder
