+++ starwarswiki/src/models/firebaseModel.js
@@ -1,5 +1,7 @@
-import { initializeApp, getDatabase, ref, get, set, onValue } from 'firebase/app';
+import { initializeApp } from 'firebase/app';
+import { getDatabase } from 'firebase/database';
 import { getAuth, GoogleAuthProvider, onAuthStateChanged, signInWithPopup } from 'firebase/auth';
+import { reactiveModel } from '../main';
 
 const app = initializeApp({
 	apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
@@ -13,16 +15,29 @@ const app = initializeApp({
 
 const db = getDatabase(app);
 
-export default function readFirebase(path) {
+/* export default function readFirebase(path) {
 	return get(ref(db, path));
-}
+} */
 
 const provider = new GoogleAuthProvider();
 provider.setCustomParameters({
 	prompt: 'select_account ',
 });
+
 export const auth = getAuth();
 
 export function signInWithGooglePopup() {
 	return signInWithPopup(auth, provider);
 }
+
+onAuthStateChanged(auth, (user) => {
+	if (user) {
+		// User is signed in
+		reactiveModel.setUser(auth.currentUser);
+		console.log('User is authorized:', user.uid);
+	} else {
+		// No user is signed in
+		reactiveModel.setUser(undefined);
+		console.log('User is not authorized');
+	}
+});
